Use Named interface in UI
Changes:
- TabPanelCareer uses the NamedListCellRenderer by hold JobType & RoleType directly in ComboBoxes
- Redesign TabPanelOrganistion to be RoleType driven instead of explicit nodes
- Move StarRater to components package

Part of #1582
